+++ internal/app/machined/pkg/controllers/secrets/kubernetes.go
@@ -19,7 +19,7 @@ import (
 	"github.com/talos-systems/crypto/x509"
 	"go.uber.org/zap"
 
-	"github.com/talos-systems/talos/internal/pkg/kubeconfig"
+	"github.com/talos-systems/talos/pkg/kubeconfig"
 	"github.com/talos-systems/talos/pkg/machinery/config"
 	"github.com/talos-systems/talos/pkg/machinery/constants"
 	"github.com/talos-systems/talos/pkg/machinery/resources/network"
